Lwm2m test OTA fix bug. Delete duplicate code.
