+++ application/src/test/java/org/thingsboard/server/transport/lwm2m/AbstractLwM2MIntegrationTest.java
@@ -1,0 +1,271 @@
+/**
+ * Copyright © 2016-2021 The Thingsboard Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.thingsboard.server.transport.lwm2m;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import org.apache.commons.io.IOUtils;
+import org.eclipse.californium.core.network.config.NetworkConfig;
+import org.eclipse.leshan.client.object.Security;
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.springframework.util.SocketUtils;
+import org.thingsboard.common.util.JacksonUtil;
+import org.thingsboard.common.util.ThingsBoardThreadFactory;
+import org.thingsboard.server.common.data.Device;
+import org.thingsboard.server.common.data.DeviceProfile;
+import org.thingsboard.server.common.data.DeviceProfileProvisionType;
+import org.thingsboard.server.common.data.DeviceProfileType;
+import org.thingsboard.server.common.data.DeviceTransportType;
+import org.thingsboard.server.common.data.ResourceType;
+import org.thingsboard.server.common.data.TbResource;
+import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MBootstrapCredentials;
+import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MClientCredentials;
+import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MDeviceCredentials;
+import org.thingsboard.server.common.data.device.credentials.lwm2m.NoSecClientCredentials;
+import org.thingsboard.server.common.data.device.credentials.lwm2m.NoSecServerCredentials;
+import org.thingsboard.server.common.data.device.profile.DefaultDeviceProfileConfiguration;
+import org.thingsboard.server.common.data.device.profile.DeviceProfileData;
+import org.thingsboard.server.common.data.device.profile.DisabledDeviceProfileProvisionConfiguration;
+import org.thingsboard.server.common.data.device.profile.Lwm2mDeviceProfileTransportConfiguration;
+import org.thingsboard.server.common.data.query.EntityData;
+import org.thingsboard.server.common.data.query.EntityDataPageLink;
+import org.thingsboard.server.common.data.query.EntityDataQuery;
+import org.thingsboard.server.common.data.query.EntityKey;
+import org.thingsboard.server.common.data.query.EntityKeyType;
+import org.thingsboard.server.common.data.query.SingleEntityFilter;
+import org.thingsboard.server.common.data.security.DeviceCredentials;
+import org.thingsboard.server.common.data.security.DeviceCredentialsType;
+import org.thingsboard.server.controller.AbstractWebsocketTest;
+import org.thingsboard.server.controller.TbTestWebSocketClient;
+import org.thingsboard.server.dao.service.DaoSqlTest;
+import org.thingsboard.server.service.telemetry.cmd.TelemetryPluginCmdsWrapper;
+import org.thingsboard.server.service.telemetry.cmd.v2.EntityDataCmd;
+import org.thingsboard.server.service.telemetry.cmd.v2.EntityDataUpdate;
+import org.thingsboard.server.service.telemetry.cmd.v2.LatestValueCmd;
+import org.thingsboard.server.transport.lwm2m.client.LwM2MTestClient;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@DaoSqlTest
+public abstract class AbstractLwM2MIntegrationTest extends AbstractWebsocketTest {
+
+    protected String transportConfiguration = "{\n" +
+            "  \"type\": \"LWM2M\",\n" +
+            "  \"observeAttr\": {\n" +
+            "    \"keyName\": {\n" +
+            "      \"/3_1.0/0/9\": \"batteryLevel\"\n" +
+            "    },\n" +
+            "    \"observe\": [],\n" +
+            "    \"attribute\": [\n" +
+            "    ],\n" +
+            "    \"telemetry\": [\n" +
+            "      \"/3_1.0/0/9\"\n" +
+            "    ],\n" +
+            "    \"attributeLwm2m\": {}\n" +
+            "  },\n" +
+            "  \"bootstrap\": {\n" +
+            "    \"servers\": {\n" +
+            "      \"binding\": \"U\",\n" +
+            "      \"shortId\": 123,\n" +
+            "      \"lifetime\": 300,\n" +
+            "      \"notifIfDisabled\": true,\n" +
+            "      \"defaultMinPeriod\": 1\n" +
+            "    },\n" +
+            "    \"lwm2mServer\": {\n" +
+            "      \"host\": \"localhost\",\n" +
+            "      \"port\": 5686,\n" +
+            "      \"serverId\": 123,\n" +
+            "      \"serverPublicKey\": \"\",\n" +
+            "      \"bootstrapServerIs\": false,\n" +
+            "      \"clientHoldOffTime\": 1,\n" +
+            "      \"bootstrapServerAccountTimeout\": 0\n" +
+            "    },\n" +
+            "    \"bootstrapServer\": {\n" +
+            "      \"host\": \"localhost\",\n" +
+            "      \"port\": 5687,\n" +
+            "      \"serverId\": 111,\n" +
+            "      \"securityMode\": \"NO_SEC\",\n" +
+            "      \"serverPublicKey\": \"\",\n" +
+            "      \"bootstrapServerIs\": true,\n" +
+            "      \"clientHoldOffTime\": 1,\n" +
+            "      \"bootstrapServerAccountTimeout\": 0\n" +
+            "    }\n" +
+            "  },\n" +
+            "  \"clientLwM2mSettings\": {\n" +
+            "    \"clientOnlyObserveAfterConnect\": 1,\n" +
+            "    \"fwUpdateStrategy\": 1,\n" +
+            "    \"swUpdateStrategy\": 1\n" +
+            "  }\n" +
+            "}";
+
+    protected DeviceProfile deviceProfile;
+    protected ScheduledExecutorService executor;
+    protected TbTestWebSocketClient wsClient;
+    protected LwM2MTestClient client;
+    private final LwM2MBootstrapCredentials defaultBootstrapCredentials;
+    private String[] resources;
+    private String endpoint;
+
+    public AbstractLwM2MIntegrationTest() {
+        this.defaultBootstrapCredentials = new LwM2MBootstrapCredentials();
+        NoSecServerCredentials serverCredentials = new NoSecServerCredentials();
+        this.defaultBootstrapCredentials.setBootstrapServer(serverCredentials);
+        this.defaultBootstrapCredentials.setLwm2mServer(serverCredentials);
+    }
+
+    public void init ()  throws Exception{
+        executor = Executors.newScheduledThreadPool(10, ThingsBoardThreadFactory.forName("test-lwm2m-scheduled"));
+        loginTenantAdmin();
+
+        for (String resourceName : this.resources) {
+            TbResource lwModel = new TbResource();
+            lwModel.setResourceType(ResourceType.LWM2M_MODEL);
+            lwModel.setTitle(resourceName);
+            lwModel.setFileName(resourceName);
+            lwModel.setTenantId(tenantId);
+            byte[] bytes = IOUtils.toByteArray(AbstractLwM2MIntegrationTest.class.getClassLoader().getResourceAsStream("lwm2m/" + resourceName));
+            lwModel.setData(Base64.getEncoder().encodeToString(bytes));
+            lwModel = doPostWithTypedResponse("/api/resource", lwModel, new TypeReference<>() {
+            });
+            Assert.assertNotNull(lwModel);
+        }
+        wsClient = buildAndConnectWebSocketClient();
+    }
+
+    @Before
+    public void beforeTest() throws Exception {
+        this.init();
+    }
+
+    @After
+    public void after() {
+        executor.shutdownNow();
+        wsClient.close();
+        clientDestroy();
+    }
+
+    public void basicTestConnectionObserveTelemetry(Security security,
+                                                    LwM2MClientCredentials credentials,
+                                                    NetworkConfig coapConfig,
+                                                    String endpoint) throws Exception {
+            createDeviceProfile(transportConfiguration);
+            Device device = createDevice(credentials);
+
+            SingleEntityFilter sef = new SingleEntityFilter();
+            sef.setSingleEntity(device.getId());
+            LatestValueCmd latestCmd = new LatestValueCmd();
+            latestCmd.setKeys(Collections.singletonList(new EntityKey(EntityKeyType.TIME_SERIES, "batteryLevel")));
+            EntityDataQuery edq = new EntityDataQuery(sef, new EntityDataPageLink(1, 0, null, null),
+                    Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+
+            EntityDataCmd cmd = new EntityDataCmd(1, edq, null, latestCmd, null);
+            TelemetryPluginCmdsWrapper wrapper = new TelemetryPluginCmdsWrapper();
+            wrapper.setEntityDataCmds(Collections.singletonList(cmd));
+
+            wsClient.send(mapper.writeValueAsString(wrapper));
+            wsClient.waitForReply();
+
+            wsClient.registerWaitForUpdate();
+            this.endpoint = endpoint;
+            createNewClient (security, coapConfig);
+            String msg = wsClient.waitForUpdate();
+
+            EntityDataUpdate update = mapper.readValue(msg, EntityDataUpdate.class);
+            Assert.assertEquals(1, update.getCmdId());
+            List<EntityData> eData = update.getUpdate();
+            Assert.assertNotNull(eData);
+            Assert.assertEquals(1, eData.size());
+            Assert.assertEquals(device.getId(), eData.get(0).getEntityId());
+            Assert.assertNotNull(eData.get(0).getLatest().get(EntityKeyType.TIME_SERIES));
+            var tsValue = eData.get(0).getLatest().get(EntityKeyType.TIME_SERIES).get("batteryLevel");
+            Assert.assertEquals(42, Long.parseLong(tsValue.getValue()));
+    }
+
+    protected void createDeviceProfile(String transportConfiguration) throws Exception {
+        deviceProfile = new DeviceProfile();
+        deviceProfile.setName("LwM2M");
+        deviceProfile.setType(DeviceProfileType.DEFAULT);
+        deviceProfile.setTenantId(tenantId);
+        deviceProfile.setTransportType(DeviceTransportType.LWM2M);
+        deviceProfile.setProvisionType(DeviceProfileProvisionType.DISABLED);
+        deviceProfile.setDescription(deviceProfile.getName());
+
+        DeviceProfileData deviceProfileData = new DeviceProfileData();
+        deviceProfileData.setConfiguration(new DefaultDeviceProfileConfiguration());
+        deviceProfileData.setProvisionConfiguration(new DisabledDeviceProfileProvisionConfiguration(null));
+        deviceProfileData.setTransportConfiguration(JacksonUtil.fromString(transportConfiguration, Lwm2mDeviceProfileTransportConfiguration.class));
+        deviceProfile.setProfileData(deviceProfileData);
+
+        deviceProfile = doPost("/api/deviceProfile", deviceProfile, DeviceProfile.class);
+        Assert.assertNotNull(deviceProfile);
+    }
+
+    protected Device createDevice(LwM2MClientCredentials clientCredentials) throws Exception {
+        Device device = new Device();
+        device.setName("Device A");
+        device.setDeviceProfileId(deviceProfile.getId());
+        device.setTenantId(tenantId);
+        device = doPost("/api/device", device, Device.class);
+        Assert.assertNotNull(device);
+
+        DeviceCredentials deviceCredentials =
+                doGet("/api/device/" + device.getId().getId().toString() + "/credentials", DeviceCredentials.class);
+        Assert.assertEquals(device.getId(), deviceCredentials.getDeviceId());
+        deviceCredentials.setCredentialsType(DeviceCredentialsType.LWM2M_CREDENTIALS);
+
+        LwM2MDeviceCredentials credentials = new LwM2MDeviceCredentials();
+        credentials.setClient(clientCredentials);
+        credentials.setBootstrap(defaultBootstrapCredentials);
+
+        deviceCredentials.setCredentialsValue(JacksonUtil.toString(credentials));
+        doPost("/api/device/credentials", deviceCredentials).andExpect(status().isOk());
+        Thread.sleep(1000);
+        return device;
+    }
+
+    public NoSecClientCredentials createNoSecClientCredentials(String endpoint) {
+        NoSecClientCredentials clientCredentials = new NoSecClientCredentials();
+        clientCredentials.setEndpoint(endpoint);
+        return clientCredentials;
+    }
+
+    public void setResources (String[] resources) {
+        this.resources = resources;
+    }
+
+    public void setEndpoint (String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public void createNewClient (Security security, NetworkConfig coapConfig)  throws Exception {
+        clientDestroy();
+        client = new LwM2MTestClient(this.executor, this.endpoint);
+        int clientPort = SocketUtils.findAvailableTcpPort();
+        client.init(security, coapConfig, clientPort);
+    }
+
+    private void clientDestroy() {
+        if (client != null) {
+            client.destroy();
+        }
+    }
+}
