+++ application/src/test/java/org/thingsboard/server/transport/lwm2m/ota/sql/OtaLwM2MIntegrationTest.java
@@ -13,21 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.transport.lwm2m.security.sql;
+package org.thingsboard.server.transport.lwm2m.ota.sql;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Assert;
 import org.junit.Test;
-import org.springframework.util.SocketUtils;
 import org.thingsboard.server.common.data.Device;
 import org.thingsboard.server.common.data.device.credentials.lwm2m.NoSecClientCredentials;
 import org.thingsboard.server.common.data.kv.KvEntry;
 import org.thingsboard.server.common.data.kv.TsKvEntry;
-import org.thingsboard.server.common.data.ota.OtaPackageType;
 import org.thingsboard.server.common.data.ota.OtaPackageUpdateStatus;
-import org.thingsboard.server.transport.lwm2m.client.LwM2MTestClient;
-import org.thingsboard.server.transport.lwm2m.security.AbstractLwM2MIntegrationTest;
+import org.thingsboard.server.transport.lwm2m.ota.AbstractOtaLwM2MIntegrationTest;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -41,9 +38,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.hamcrest.Matchers.hasSize;
 import static org.thingsboard.rest.client.utils.RestJsonConverter.toTimeseries;
-import static org.thingsboard.server.common.data.ota.OtaPackageKey.CHECKSUM;
-import static org.thingsboard.server.common.data.ota.OtaPackageKey.CHECKSUM_ALGORITHM;
-import static org.thingsboard.server.common.data.ota.OtaPackageKey.SIZE;
 import static org.thingsboard.server.common.data.ota.OtaPackageUpdateStatus.DOWNLOADED;
 import static org.thingsboard.server.common.data.ota.OtaPackageUpdateStatus.DOWNLOADING;
 import static org.thingsboard.server.common.data.ota.OtaPackageUpdateStatus.FAILED;
@@ -56,7 +50,7 @@ import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.COAP_CONFIG
 import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.SECURITY;
 
 @Slf4j
-public class OtaLwM2MIntegrationTest extends AbstractLwM2MIntegrationTest {
+public class OtaLwM2MIntegrationTest extends AbstractOtaLwM2MIntegrationTest {
 
     public static final int TIMEOUT = 30;
     private final String OTA_TRANSPORT_CONFIGURATION = "{\n" +
@@ -139,56 +133,60 @@ public class OtaLwM2MIntegrationTest extends AbstractLwM2MIntegrationTest {
     @Test
     public void testFirmwareUpdateWithClientWithoutFirmwareOtaInfoFromProfile() throws Exception {
         String endpoint = "WithoutFirmwareInfoDevice";
-            createDeviceProfile(TRANSPORT_CONFIGURATION);
-            NoSecClientCredentials clientCredentials = new NoSecClientCredentials();
-            clientCredentials.setEndpoint(endpoint);
-            Device device = createDevice(clientCredentials);
-            client = new LwM2MTestClient(executor, endpoint);
-            int clientPort = SocketUtils.findAvailableTcpPort();
-            client.init(SECURITY, COAP_CONFIG, clientPort);
+        setEndpoint(endpoint);
+        createDeviceProfile(transportConfiguration);
+        NoSecClientCredentials credentials = createNoSecClientCredentials(endpoint);
+        final Device device = createDevice(credentials);
+        createNewClient(SECURITY, COAP_CONFIG);
 
-            Thread.sleep(1000);
+        Thread.sleep(1000);
 
-            device.setFirmwareId(createFirmware().getId());
-            device = doPost("/api/device", device, Device.class);
+        device.setFirmwareId(createFirmware().getId());
+        final Device savedDevice = doPost("/api/device", device, Device.class);
 
-            Thread.sleep(5000);
+        Thread.sleep(1000);
 
-            List<TsKvEntry> ts = toTimeseries(doGetAsyncTyped("/api/plugins/telemetry/DEVICE/" + device.getId().getId() + "/values/timeseries?keys=fw_state", new TypeReference<>() {
-            }));
+        assertThat(savedDevice).as("saved device").isNotNull();
+        assertThat(getDeviceFromAPI(device.getId().getId())).as("fetched device").isEqualTo(savedDevice);
 
-            List<OtaPackageUpdateStatus> statuses = ts.stream().map(KvEntry::getValueAsString).map(OtaPackageUpdateStatus::valueOf).collect(Collectors.toList());
+        List<TsKvEntry> ts = toTimeseries(doGetAsyncTyped("/api/plugins/telemetry/DEVICE/" +
+                savedDevice.getId().getId() + "/values/timeseries?keys=fw_state", new TypeReference<>() {}));
+        List<OtaPackageUpdateStatus> statuses = ts.stream().map(KvEntry::getValueAsString).map(OtaPackageUpdateStatus::valueOf).collect(Collectors.toList());
+        List<OtaPackageUpdateStatus> expectedStatuses = Collections.singletonList(FAILED);
 
-            List<OtaPackageUpdateStatus> expectedStatuses = Collections.singletonList(FAILED);
-
-            Assert.assertEquals(expectedStatuses, statuses);
+        Assert.assertEquals(expectedStatuses, statuses);
     }
 
     @Test
     public void testFirmwareUpdateByObject5() throws Exception {
         String endpoint = "Ota5_Device";
-        List<OtaPackageUpdateStatus> expectedStatuses = Arrays.asList(QUEUED, INITIATED, DOWNLOADING, DOWNLOADED, UPDATING, UPDATED);
+        setEndpoint(endpoint);
         createDeviceProfile(OTA_TRANSPORT_CONFIGURATION);
-        NoSecClientCredentials clientCredentials = new NoSecClientCredentials();
-        clientCredentials.setEndpoint(endpoint);
-        Device device = createDevice(clientCredentials);
+        NoSecClientCredentials credentials = createNoSecClientCredentials(endpoint);
+        final Device device = createDevice(credentials);
+        createNewClient(SECURITY, COAP_CONFIG);
 
-        device.setFirmwareId(createFirmware().getId());
-        Device saveDevice = doPost("/api/device", device, Device.class);
         Thread.sleep(1000);
 
-        client = new LwM2MTestClient(executor, endpoint);
-        int clientPort = SocketUtils.findAvailableTcpPort();
-        client.init(SECURITY, COAP_CONFIG, clientPort);
+        device.setFirmwareId(createFirmware().getId());
+        final Device savedDevice = doPost("/api/device", device, Device.class);
 
+        Thread.sleep(1000);
+
+        assertThat(savedDevice).as("saved device").isNotNull();
+        assertThat(getDeviceFromAPI(device.getId().getId())).as("fetched device").isEqualTo(savedDevice);
+
+        final List<OtaPackageUpdateStatus> expectedStatuses = Arrays.asList(QUEUED, INITIATED, DOWNLOADING, DOWNLOADED, UPDATING, UPDATED);
         List<TsKvEntry> ts = await("await on timeseries")
                 .atMost(30, TimeUnit.SECONDS)
-                .until(() -> toTimeseries(doGetAsyncTyped("/api/plugins/telemetry/DEVICE/" + saveDevice.getId().getId() + "/values/timeseries?orderBy=ASC&keys=fw_state&startTs=0&endTs=" + System.currentTimeMillis(), new TypeReference<>() {
+                .until(() -> toTimeseries(doGetAsyncTyped("/api/plugins/telemetry/DEVICE/" +
+                        savedDevice.getId().getId() + "/values/timeseries?orderBy=ASC&keys=fw_state&startTs=0&endTs=" +
+                        System.currentTimeMillis(), new TypeReference<>() {
                 })), hasSize(expectedStatuses.size()));
-
-
-        List<OtaPackageUpdateStatus> statuses = ts.stream().sorted(Comparator.comparingLong(TsKvEntry::getTs)).map(KvEntry::getValueAsString).map(OtaPackageUpdateStatus::valueOf).collect(Collectors.toList());
-
+        List<OtaPackageUpdateStatus> statuses = ts.stream().sorted(Comparator
+                        .comparingLong(TsKvEntry::getTs)).map(KvEntry::getValueAsString)
+                .map(OtaPackageUpdateStatus::valueOf)
+                .collect(Collectors.toList());
 
         Assert.assertEquals(expectedStatuses, statuses);
     }
@@ -200,28 +198,25 @@ public class OtaLwM2MIntegrationTest extends AbstractLwM2MIntegrationTest {
      * */
     @Test
     public void testSoftwareUpdateByObject9() throws Exception {
-        //given
-        final List<OtaPackageUpdateStatus> expectedStatuses = List.of(
-                QUEUED, INITIATED, DOWNLOADING, DOWNLOADING, DOWNLOADING, DOWNLOADED, VERIFIED, UPDATED);
         String endpoint = "Ota9_Device";
+        setEndpoint(endpoint);
         createDeviceProfile(OTA_TRANSPORT_CONFIGURATION);
-        NoSecClientCredentials clientCredentials = new NoSecClientCredentials();
-        clientCredentials.setEndpoint(endpoint);
-        final Device device = createDevice(clientCredentials);
-        device.setSoftwareId(createSoftware().getId());
+        NoSecClientCredentials credentials = createNoSecClientCredentials(endpoint);
+        final Device device = createDevice(credentials);
+        createNewClient(SECURITY, COAP_CONFIG);
 
+        Thread.sleep(1000);
+
+        device.setSoftwareId(createSoftware().getId());
         final Device savedDevice = doPost("/api/device", device, Device.class); //sync call
+
         Thread.sleep(1000);
 
         assertThat(savedDevice).as("saved device").isNotNull();
         assertThat(getDeviceFromAPI(device.getId().getId())).as("fetched device").isEqualTo(savedDevice);
 
-        //when
-        log.warn("Init the client...");
-        client = new LwM2MTestClient(executor, endpoint);
-        int clientPort = SocketUtils.findAvailableTcpPort();
-        client.init(SECURITY, COAP_CONFIG, clientPort);
-
+        final List<OtaPackageUpdateStatus> expectedStatuses = List.of(
+                QUEUED, INITIATED, DOWNLOADING, DOWNLOADING, DOWNLOADING, DOWNLOADED, VERIFIED, UPDATED);
         log.warn("AWAIT atMost {} SECONDS on timeseries List<TsKvEntry> by API with list size {}...", TIMEOUT, expectedStatuses.size());
         List<TsKvEntry> ts = await("await on timeseries")
                 .atMost(30, TimeUnit.SECONDS)
@@ -230,9 +225,10 @@ public class OtaLwM2MIntegrationTest extends AbstractLwM2MIntegrationTest {
 
         ts.sort(Comparator.comparingLong(TsKvEntry::getTs));
         log.warn("Ts ordered: {}", ts);
-        ts.forEach((x) -> log.warn("ts: {} ", x));
-
-        List<OtaPackageUpdateStatus> statuses = ts.stream().map(KvEntry::getValueAsString).map(OtaPackageUpdateStatus::valueOf).collect(Collectors.toList());
+        ts.forEach((x) -> log.warn("ts: {        Thread.sleep(1000);} ", x));
+        List<OtaPackageUpdateStatus> statuses = ts.stream().map(KvEntry::getValueAsString)
+                .map(OtaPackageUpdateStatus::valueOf)
+                .collect(Collectors.toList());
         log.warn("Converted ts to statuses: {}", statuses);
 
         assertThat(statuses).isEqualTo(expectedStatuses);
