--- application/src/test/java/org/thingsboard/server/transport/lwm2m/security/AbstractLwM2MIntegrationTest.java
@@ -1,430 +1,0 @@
-/**
- * Copyright © 2016-2021 The Thingsboard Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.thingsboard.server.transport.lwm2m.security;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import org.apache.commons.io.IOUtils;
-import org.eclipse.californium.core.CoapClient;
-import org.eclipse.californium.core.network.config.NetworkConfig;
-import org.eclipse.leshan.client.object.Security;
-import org.eclipse.leshan.core.util.Hex;
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.test.web.servlet.request.MockMultipartHttpServletRequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.util.SocketUtils;
-import org.thingsboard.common.util.JacksonUtil;
-import org.thingsboard.common.util.ThingsBoardThreadFactory;
-import org.thingsboard.server.common.data.Device;
-import org.thingsboard.server.common.data.DeviceProfile;
-import org.thingsboard.server.common.data.DeviceProfileProvisionType;
-import org.thingsboard.server.common.data.DeviceProfileType;
-import org.thingsboard.server.common.data.DeviceTransportType;
-import org.thingsboard.server.common.data.OtaPackageInfo;
-import org.thingsboard.server.common.data.ResourceType;
-import org.thingsboard.server.common.data.TbResource;
-import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MBootstrapCredentials;
-import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MClientCredentials;
-import org.thingsboard.server.common.data.device.credentials.lwm2m.LwM2MDeviceCredentials;
-import org.thingsboard.server.common.data.device.credentials.lwm2m.NoSecServerCredentials;
-import org.thingsboard.server.common.data.device.profile.DefaultDeviceProfileConfiguration;
-import org.thingsboard.server.common.data.device.profile.DeviceProfileData;
-import org.thingsboard.server.common.data.device.profile.DisabledDeviceProfileProvisionConfiguration;
-import org.thingsboard.server.common.data.device.profile.Lwm2mDeviceProfileTransportConfiguration;
-import org.thingsboard.server.common.data.query.EntityData;
-import org.thingsboard.server.common.data.query.EntityDataPageLink;
-import org.thingsboard.server.common.data.query.EntityDataQuery;
-import org.thingsboard.server.common.data.query.EntityKey;
-import org.thingsboard.server.common.data.query.EntityKeyType;
-import org.thingsboard.server.common.data.query.SingleEntityFilter;
-import org.thingsboard.server.common.data.security.DeviceCredentials;
-import org.thingsboard.server.common.data.security.DeviceCredentialsType;
-import org.thingsboard.server.controller.AbstractWebsocketTest;
-import org.thingsboard.server.controller.TbTestWebSocketClient;
-import org.thingsboard.server.dao.service.DaoSqlTest;
-import org.thingsboard.server.service.telemetry.cmd.TelemetryPluginCmdsWrapper;
-import org.thingsboard.server.service.telemetry.cmd.v2.EntityDataCmd;
-import org.thingsboard.server.service.telemetry.cmd.v2.EntityDataUpdate;
-import org.thingsboard.server.service.telemetry.cmd.v2.LatestValueCmd;
-import org.thingsboard.server.transport.lwm2m.client.LwM2MTestClient;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.math.BigInteger;
-import java.security.AlgorithmParameters;
-import java.security.GeneralSecurityException;
-import java.security.KeyFactory;
-import java.security.KeyStore;
-import java.security.PrivateKey;
-import java.security.PublicKey;
-import java.security.cert.Certificate;
-import java.security.cert.X509Certificate;
-import java.security.spec.ECGenParameterSpec;
-import java.security.spec.ECParameterSpec;
-import java.security.spec.ECPoint;
-import java.security.spec.ECPrivateKeySpec;
-import java.security.spec.ECPublicKeySpec;
-import java.security.spec.KeySpec;
-import java.util.Base64;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-
-import static org.eclipse.leshan.client.object.Security.noSec;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.thingsboard.server.common.data.ota.OtaPackageType.FIRMWARE;
-import static org.thingsboard.server.common.data.ota.OtaPackageType.SOFTWARE;
-import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.COAP_CONFIG;
-import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.SECURITY;
-
-@DaoSqlTest
-public abstract class AbstractLwM2MIntegrationTest extends AbstractWebsocketTest {
-
-    protected final String TRANSPORT_CONFIGURATION = "{\n" +
-            "  \"type\": \"LWM2M\",\n" +
-            "  \"observeAttr\": {\n" +
-            "    \"keyName\": {\n" +
-            "      \"/3_1.0/0/9\": \"batteryLevel\"\n" +
-            "    },\n" +
-            "    \"observe\": [],\n" +
-            "    \"attribute\": [\n" +
-            "    ],\n" +
-            "    \"telemetry\": [\n" +
-            "      \"/3_1.0/0/9\"\n" +
-            "    ],\n" +
-            "    \"attributeLwm2m\": {}\n" +
-            "  },\n" +
-            "  \"bootstrap\": {\n" +
-            "    \"servers\": {\n" +
-            "      \"binding\": \"U\",\n" +
-            "      \"shortId\": 123,\n" +
-            "      \"lifetime\": 300,\n" +
-            "      \"notifIfDisabled\": true,\n" +
-            "      \"defaultMinPeriod\": 1\n" +
-            "    },\n" +
-            "    \"lwm2mServer\": {\n" +
-            "      \"host\": \"localhost\",\n" +
-            "      \"port\": 5686,\n" +
-            "      \"serverId\": 123,\n" +
-            "      \"serverPublicKey\": \"\",\n" +
-            "      \"bootstrapServerIs\": false,\n" +
-            "      \"clientHoldOffTime\": 1,\n" +
-            "      \"bootstrapServerAccountTimeout\": 0\n" +
-            "    },\n" +
-            "    \"bootstrapServer\": {\n" +
-            "      \"host\": \"localhost\",\n" +
-            "      \"port\": 5687,\n" +
-            "      \"serverId\": 111,\n" +
-            "      \"securityMode\": \"NO_SEC\",\n" +
-            "      \"serverPublicKey\": \"\",\n" +
-            "      \"bootstrapServerIs\": true,\n" +
-            "      \"clientHoldOffTime\": 1,\n" +
-            "      \"bootstrapServerAccountTimeout\": 0\n" +
-            "    }\n" +
-            "  },\n" +
-            "  \"clientLwM2mSettings\": {\n" +
-            "    \"clientOnlyObserveAfterConnect\": 1,\n" +
-            "    \"fwUpdateStrategy\": 1,\n" +
-            "    \"swUpdateStrategy\": 1\n" +
-            "  }\n" +
-            "}";
-
-    protected DeviceProfile deviceProfile;
-    protected ScheduledExecutorService executor;
-    protected TbTestWebSocketClient wsClient;
-
-    protected final PublicKey clientPublicKey; // client public key used for RPK
-    protected final PrivateKey clientPrivateKey; // client private key used for RPK
-    protected final PublicKey serverPublicKey; // server public key used for RPK
-    protected final PrivateKey serverPrivateKey; // server private key used for RPK
-
-    // client private key used for X509
-    protected final PrivateKey clientPrivateKeyFromCert;
-    // server private key used for X509
-    protected final PrivateKey serverPrivateKeyFromCert;
-    // client certificate signed by rootCA with a good CN (CN start by leshan_integration_test)
-    protected final X509Certificate clientX509Cert;
-    // client certificate signed by rootCA but with bad CN (CN does not start by leshan_integration_test)
-    protected final X509Certificate clientX509CertWithBadCN;
-    // client certificate self-signed with a good CN (CN start by leshan_integration_test)
-    protected final X509Certificate clientX509CertSelfSigned;
-    // client certificate signed by another CA (not rootCA) with a good CN (CN start by leshan_integration_test)
-    protected final X509Certificate clientX509CertNotTrusted;
-    // server certificate signed by rootCA
-    protected final X509Certificate serverX509Cert;
-    // self-signed server certificate
-    protected final X509Certificate serverX509CertSelfSigned;
-    // rootCA used by the server
-    protected final X509Certificate rootCAX509Cert;
-    // certificates trustedby the server (should contain rootCA)
-    protected final Certificate[] trustedCertificates = new Certificate[1];
-
-    protected static final String ENDPOINT = "deviceAEndpoint";
-
-    protected LwM2MTestClient client;
-
-    private final LwM2MBootstrapCredentials defaultBootstrapCredentials;
-
-    public AbstractLwM2MIntegrationTest() {
-// create client credentials
-        try {
-            // Get point values
-            byte[] publicX = Hex
-                    .decodeHex("89c048261979208666f2bfb188be1968fc9021c416ce12828c06f4e314c167b5".toCharArray());
-            byte[] publicY = Hex
-                    .decodeHex("cbf1eb7587f08e01688d9ada4be859137ca49f79394bad9179326b3090967b68".toCharArray());
-            byte[] privateS = Hex
-                    .decodeHex("e67b68d2aaeb6550f19d98cade3ad62b39532e02e6b422e1f7ea189dabaea5d2".toCharArray());
-
-            // Get Elliptic Curve Parameter spec for secp256r1
-            AlgorithmParameters algoParameters = AlgorithmParameters.getInstance("EC");
-            algoParameters.init(new ECGenParameterSpec("secp256r1"));
-            ECParameterSpec parameterSpec = algoParameters.getParameterSpec(ECParameterSpec.class);
-
-            // Create key specs
-            KeySpec publicKeySpec = new ECPublicKeySpec(new ECPoint(new BigInteger(publicX), new BigInteger(publicY)),
-                    parameterSpec);
-            KeySpec privateKeySpec = new ECPrivateKeySpec(new BigInteger(privateS), parameterSpec);
-
-            // Get keys
-            clientPublicKey = KeyFactory.getInstance("EC").generatePublic(publicKeySpec);
-            clientPrivateKey = KeyFactory.getInstance("EC").generatePrivate(privateKeySpec);
-
-            // Get certificates from key store
-            char[] clientKeyStorePwd = "client".toCharArray();
-            KeyStore clientKeyStore = KeyStore.getInstance(KeyStore.getDefaultType());
-            try (InputStream clientKeyStoreFile = this.getClass().getClassLoader().getResourceAsStream("lwm2m/credentials/clientKeyStore.jks")) {
-                clientKeyStore.load(clientKeyStoreFile, clientKeyStorePwd);
-            }
-
-            clientPrivateKeyFromCert = (PrivateKey) clientKeyStore.getKey("client", clientKeyStorePwd);
-            clientX509Cert = (X509Certificate) clientKeyStore.getCertificate("client");
-            clientX509CertWithBadCN = (X509Certificate) clientKeyStore.getCertificate("client_bad_cn");
-            clientX509CertSelfSigned = (X509Certificate) clientKeyStore.getCertificate("client_self_signed");
-            clientX509CertNotTrusted = (X509Certificate) clientKeyStore.getCertificate("client_not_trusted");
-        } catch (GeneralSecurityException | IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        // create server credentials
-        try {
-            // Get point values
-            byte[] publicX = Hex
-                    .decodeHex("fcc28728c123b155be410fc1c0651da374fc6ebe7f96606e90d927d188894a73".toCharArray());
-            byte[] publicY = Hex
-                    .decodeHex("d2ffaa73957d76984633fc1cc54d0b763ca0559a9dff9706e9f4557dacc3f52a".toCharArray());
-            byte[] privateS = Hex
-                    .decodeHex("1dae121ba406802ef07c193c1ee4df91115aabd79c1ed7f4c0ef7ef6a5449400".toCharArray());
-
-            // Get Elliptic Curve Parameter spec for secp256r1
-            AlgorithmParameters algoParameters = AlgorithmParameters.getInstance("EC");
-            algoParameters.init(new ECGenParameterSpec("secp256r1"));
-            ECParameterSpec parameterSpec = algoParameters.getParameterSpec(ECParameterSpec.class);
-
-            // Create key specs
-            KeySpec publicKeySpec = new ECPublicKeySpec(new ECPoint(new BigInteger(publicX), new BigInteger(publicY)),
-                    parameterSpec);
-            KeySpec privateKeySpec = new ECPrivateKeySpec(new BigInteger(privateS), parameterSpec);
-
-            // Get keys
-            serverPublicKey = KeyFactory.getInstance("EC").generatePublic(publicKeySpec);
-            serverPrivateKey = KeyFactory.getInstance("EC").generatePrivate(privateKeySpec);
-
-            // Get certificates from key store
-            char[] serverKeyStorePwd = "server".toCharArray();
-            KeyStore serverKeyStore = KeyStore.getInstance(KeyStore.getDefaultType());
-            try (InputStream serverKeyStoreFile = this.getClass().getClassLoader().getResourceAsStream("lwm2m/credentials/serverKeyStore.jks")) {
-                serverKeyStore.load(serverKeyStoreFile, serverKeyStorePwd);
-            }
-
-            serverPrivateKeyFromCert = (PrivateKey) serverKeyStore.getKey("server", serverKeyStorePwd);
-            rootCAX509Cert = (X509Certificate) serverKeyStore.getCertificate("rootCA");
-            serverX509Cert = (X509Certificate) serverKeyStore.getCertificate("server");
-            serverX509CertSelfSigned = (X509Certificate) serverKeyStore.getCertificate("server_self_signed");
-            trustedCertificates[0] = rootCAX509Cert;
-        } catch (GeneralSecurityException | IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        defaultBootstrapCredentials = new LwM2MBootstrapCredentials();
-
-        NoSecServerCredentials serverCredentials = new NoSecServerCredentials();
-
-        defaultBootstrapCredentials.setBootstrapServer(serverCredentials);
-        defaultBootstrapCredentials.setLwm2mServer(serverCredentials);
-    }
-
-    @Before
-    public void beforeTest() throws Exception {
-        executor = Executors.newScheduledThreadPool(10, ThingsBoardThreadFactory.forName("test-lwm2m-scheduled"));
-        loginTenantAdmin();
-
-        String[] resources = new String[]{"1.xml", "2.xml", "3.xml", "5.xml", "9.xml"};
-        for (String resourceName : resources) {
-            TbResource lwModel = new TbResource();
-            lwModel.setResourceType(ResourceType.LWM2M_MODEL);
-            lwModel.setTitle(resourceName);
-            lwModel.setFileName(resourceName);
-            lwModel.setTenantId(tenantId);
-            byte[] bytes = IOUtils.toByteArray(AbstractLwM2MIntegrationTest.class.getClassLoader().getResourceAsStream("lwm2m/" + resourceName));
-            lwModel.setData(Base64.getEncoder().encodeToString(bytes));
-            lwModel = doPostWithTypedResponse("/api/resource", lwModel, new TypeReference<>() {
-            });
-            Assert.assertNotNull(lwModel);
-        }
-        wsClient = buildAndConnectWebSocketClient();
-    }
-
-    protected void createDeviceProfile(String transportConfiguration) throws Exception {
-        deviceProfile = new DeviceProfile();
-
-        deviceProfile.setName("LwM2M");
-        deviceProfile.setType(DeviceProfileType.DEFAULT);
-        deviceProfile.setTenantId(tenantId);
-        deviceProfile.setTransportType(DeviceTransportType.LWM2M);
-        deviceProfile.setProvisionType(DeviceProfileProvisionType.DISABLED);
-        deviceProfile.setDescription(deviceProfile.getName());
-
-        DeviceProfileData deviceProfileData = new DeviceProfileData();
-        deviceProfileData.setConfiguration(new DefaultDeviceProfileConfiguration());
-        deviceProfileData.setProvisionConfiguration(new DisabledDeviceProfileProvisionConfiguration(null));
-        deviceProfileData.setTransportConfiguration(JacksonUtil.fromString(transportConfiguration, Lwm2mDeviceProfileTransportConfiguration.class));
-        deviceProfile.setProfileData(deviceProfileData);
-
-        deviceProfile = doPost("/api/deviceProfile", deviceProfile, DeviceProfile.class);
-        Assert.assertNotNull(deviceProfile);
-    }
-
-    protected Device createDevice(LwM2MClientCredentials clientCredentials) throws Exception {
-        Device device = new Device();
-        device.setName("Device A");
-        device.setDeviceProfileId(deviceProfile.getId());
-        device.setTenantId(tenantId);
-        device = doPost("/api/device", device, Device.class);
-        Assert.assertNotNull(device);
-
-        DeviceCredentials deviceCredentials =
-                doGet("/api/device/" + device.getId().getId().toString() + "/credentials", DeviceCredentials.class);
-        Assert.assertEquals(device.getId(), deviceCredentials.getDeviceId());
-        deviceCredentials.setCredentialsType(DeviceCredentialsType.LWM2M_CREDENTIALS);
-
-        LwM2MDeviceCredentials credentials = new LwM2MDeviceCredentials();
-
-        credentials.setClient(clientCredentials);
-        credentials.setBootstrap(defaultBootstrapCredentials);
-
-        deviceCredentials.setCredentialsValue(JacksonUtil.toString(credentials));
-        doPost("/api/device/credentials", deviceCredentials).andExpect(status().isOk());
-        return device;
-    }
-
-
-    protected OtaPackageInfo createFirmware() throws Exception {
-        String CHECKSUM = "4bf5122f344554c53bde2ebb8cd2b7e3d1600ad631c385a5d7cce23c7785459a";
-
-        OtaPackageInfo firmwareInfo = new OtaPackageInfo();
-        firmwareInfo.setDeviceProfileId(deviceProfile.getId());
-        firmwareInfo.setType(FIRMWARE);
-        firmwareInfo.setTitle("My firmware");
-        firmwareInfo.setVersion("v1.0");
-
-        OtaPackageInfo savedFirmwareInfo = doPost("/api/otaPackage", firmwareInfo, OtaPackageInfo.class);
-
-        MockMultipartFile testData = new MockMultipartFile("file", "filename.txt", "text/plain", new byte[]{1});
-
-        return savaData("/api/otaPackage/" + savedFirmwareInfo.getId().getId().toString() + "?checksum={checksum}&checksumAlgorithm={checksumAlgorithm}", testData, CHECKSUM, "SHA256");
-    }
-
-    protected OtaPackageInfo createSoftware() throws Exception {
-        String CHECKSUM = "4bf5122f344554c53bde2ebb8cd2b7e3d1600ad631c385a5d7cce23c7785459a";
-
-        OtaPackageInfo swInfo = new OtaPackageInfo();
-        swInfo.setDeviceProfileId(deviceProfile.getId());
-        swInfo.setType(SOFTWARE);
-        swInfo.setTitle("My sw");
-        swInfo.setVersion("v1.0");
-
-        OtaPackageInfo savedFirmwareInfo = doPost("/api/otaPackage", swInfo, OtaPackageInfo.class);
-
-        MockMultipartFile testData = new MockMultipartFile("file", "filename.txt", "text/plain", new byte[]{1});
-
-        return savaData("/api/otaPackage/" + savedFirmwareInfo.getId().getId().toString() + "?checksum={checksum}&checksumAlgorithm={checksumAlgorithm}", testData, CHECKSUM, "SHA256");
-    }
-
-    protected OtaPackageInfo savaData(String urlTemplate, MockMultipartFile content, String... params) throws Exception {
-        MockMultipartHttpServletRequestBuilder postRequest = MockMvcRequestBuilders.multipart(urlTemplate, params);
-        postRequest.file(content);
-        setJwtToken(postRequest);
-        return readResponse(mockMvc.perform(postRequest).andExpect(status().isOk()), OtaPackageInfo.class);
-    }
-
-    @After
-    public void after() {
-        executor.shutdownNow();
-        wsClient.close();
-        if (client != null) {
-            client.destroy();
-        }
-    }
-
-    public void basicTestConnectionObserveTelemetry(Security security,
-                                                    LwM2MClientCredentials credentials,
-                                                    NetworkConfig coapConfig,
-                                                    String endpoint) throws Exception {
-        LwM2MTestClient client = null;
-        try {
-            createDeviceProfile(TRANSPORT_CONFIGURATION);
-            Device device = createDevice(credentials);
-
-            SingleEntityFilter sef = new SingleEntityFilter();
-            sef.setSingleEntity(device.getId());
-            LatestValueCmd latestCmd = new LatestValueCmd();
-            latestCmd.setKeys(Collections.singletonList(new EntityKey(EntityKeyType.TIME_SERIES, "batteryLevel")));
-            EntityDataQuery edq = new EntityDataQuery(sef, new EntityDataPageLink(1, 0, null, null),
-                    Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-
-            EntityDataCmd cmd = new EntityDataCmd(1, edq, null, latestCmd, null);
-            TelemetryPluginCmdsWrapper wrapper = new TelemetryPluginCmdsWrapper();
-            wrapper.setEntityDataCmds(Collections.singletonList(cmd));
-
-            wsClient.send(mapper.writeValueAsString(wrapper));
-            wsClient.waitForReply();
-
-            wsClient.registerWaitForUpdate();
-            client = new LwM2MTestClient(executor, endpoint);
-            int clientPort = SocketUtils.findAvailableTcpPort();
-            client.init(security, coapConfig, clientPort);
-            String msg = wsClient.waitForUpdate();
-
-            EntityDataUpdate update = mapper.readValue(msg, EntityDataUpdate.class);
-            Assert.assertEquals(1, update.getCmdId());
-            List<EntityData> eData = update.getUpdate();
-            Assert.assertNotNull(eData);
-            Assert.assertEquals(1, eData.size());
-            Assert.assertEquals(device.getId(), eData.get(0).getEntityId());
-            Assert.assertNotNull(eData.get(0).getLatest().get(EntityKeyType.TIME_SERIES));
-            var tsValue = eData.get(0).getLatest().get(EntityKeyType.TIME_SERIES).get("batteryLevel");
-            Assert.assertEquals(42, Long.parseLong(tsValue.getValue()));
-        } finally {
-            if (client != null) {
-                client.destroy();
-            }
-        }
-    }
-}
