+++ application/src/test/java/org/thingsboard/server/transport/lwm2m/security/sql/X509_NoTrustLwM2MIntegrationTest.java
@@ -19,14 +19,14 @@ import org.eclipse.leshan.client.object.Security;
 import org.junit.Test;
 import org.thingsboard.server.common.data.device.credentials.lwm2m.X509ClientCredentials;
 import org.thingsboard.server.common.transport.util.SslUtil;
-import org.thingsboard.server.transport.lwm2m.security.AbstractLwM2MIntegrationTest;
+import org.thingsboard.server.transport.lwm2m.security.AbstractSecurityLwM2MIntegrationTest;
 
 import static org.eclipse.leshan.client.object.Security.x509;
 import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.SECURE_COAP_CONFIG;
 import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.SECURE_URI;
 import static org.thingsboard.server.transport.lwm2m.Lwm2mTestHelper.SHORT_SERVER_ID;
 
-public class X509_NoTrustLwM2MIntegrationTest extends AbstractLwM2MIntegrationTest {
+public class X509_NoTrustLwM2MIntegrationTest extends AbstractSecurityLwM2MIntegrationTest {
 
     @Test
     public void testConnectWithCertAndObserveTelemetry() throws Exception {
@@ -34,7 +34,7 @@ public class X509_NoTrustLwM2MIntegrationTest extends AbstractLwM2MIntegrationTe
         credentials.setEndpoint(ENDPOINT);
         credentials.setCert(SslUtil.getCertificateString(clientX509CertNotTrusted));
         Security security = x509(SECURE_URI,
-                123,
+                SHORT_SERVER_ID,
                 clientX509CertNotTrusted.getEncoded(),
                 clientPrivateKeyFromCert.getEncoded(),
                 serverX509Cert.getEncoded());
